backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]>=0.31.1
gunicorn>=21.2.0
python-dotenv==1.0.1
pydantic>=2,<3
httpx>=0.27.1
azure-search-documents==11.6.0b3
openai>=2.0.0
agent-framework-core==1.0.0rc3
azure-cosmos>=4.7.0
azure-identity>=1.17.0
